Add validation when delete teacher data

+++ src/teachers/teachers.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { TeacherRepository } from './repository/teacher.repository';
+import { ClassRepository } from '../classes/repository/class.repository'
 import { Teacher } from './entity/teacher.entity';
 import { CreateTeacherDto } from './dto/create-teacher.dto';
 import { UpdateTeacherDto } from './dto/update-teacher.dto';
@@ -8,7 +9,10 @@ import { ErrorResponse } from "./response/error-response.interface"
 @Injectable()
 export class TeachersService {
 
-    constructor(private readonly teacherRepository: TeacherRepository) {}
+    constructor(
+        private readonly teacherRepository: TeacherRepository,
+        private readonly classRepository: ClassRepository,
+    ) {}
 
     async getDatas(): Promise<Teacher[]> {
         try {
@@ -41,10 +45,17 @@ export class TeachersService {
 
     async deleteData(id: string): Promise<void> {
         try {
+
             const data = await this.teacherRepository.findOneBy({'teacher_id': id})
             if (!data) {
                 throw new ErrorResponse("Data not found")
             }
+
+            const countClassData = await this.classRepository.countBy({'reference_teacher_id': id})
+            if (countClassData > 0) {
+                throw new ErrorResponse("Cannot delete subject data, already used in class")
+            }
+
             await this.teacherRepository.delete({teacher_id: id})
         } catch(e) {
             throw new ErrorResponse(e.message)
